Fix bug

+++ tests/unit/linalg/test_ode.cpp
@@ -184,70 +184,70 @@ TEST_CASE("First order ODE methods",
    SECTION("BackwardEuler")
    {
       std::cout <<"\nTesting BackwardEuler" << std::endl;
-      int conv_rate = check.order(new BackwardEulerSolver);
+      double conv_rate = check.order(new BackwardEulerSolver);
       REQUIRE(conv_rate + tol > 1.0);
    }
 
    SECTION("SDIRK23Solver(2)")
    {
       std::cout <<"\nTesting SDIRK23Solver(2)" << std::endl;
-      int conv_rate = check.order(new SDIRK23Solver(2));
+      double conv_rate = check.order(new SDIRK23Solver(2));
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("SDIRK33Solver")
    {
       std::cout <<"\nTesting SDIRK33Solver" << std::endl;
-      int conv_rate = check.order(new SDIRK33Solver);
+      double conv_rate = check.order(new SDIRK33Solver);
       REQUIRE(conv_rate + tol > 3.0);
    }
 
    SECTION("ForwardEulerSolver")
    {
       std::cout <<"\nTesting ForwardEulerSolver" << std::endl;
-      int conv_rate = check.order(new ForwardEulerSolver);
+      double conv_rate = check.order(new ForwardEulerSolver);
       REQUIRE(conv_rate + tol > 1.0);
    }
 
    SECTION("RK2Solver(0.5)")
    {
       std::cout <<"\nTesting RK2Solver(0.5)" << std::endl;
-      int conv_rate = check.order(new RK2Solver(0.5));
+      double conv_rate = check.order(new RK2Solver(0.5));
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("RK3SSPSolver")
    {
       std::cout <<"\nTesting RK3SSPSolver" << std::endl;
-      int conv_rate = check.order(new RK3SSPSolver);
+      double conv_rate = check.order(new RK3SSPSolver);
       REQUIRE(conv_rate + tol > 3.0);
    }
 
    SECTION("RK4Solver")
    {
       std::cout <<"\nTesting RK4Solver" << std::endl;
-      int conv_rate = check.order(new RK4Solver);
+      double conv_rate = check.order(new RK4Solver);
       REQUIRE(conv_rate + tol > 4.0);
    }
 
    SECTION("ImplicitMidpointSolver")
    {
       std::cout <<"\nTesting ImplicitMidpointSolver" << std::endl;
-      int conv_rate = check.order(new ImplicitMidpointSolver);
+      double conv_rate = check.order(new ImplicitMidpointSolver);
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("SDIRK23Solver")
    {
       std::cout <<"\nTesting SDIRK23Solver" << std::endl;
-      int conv_rate = check.order(new SDIRK23Solver);
+      double conv_rate = check.order(new SDIRK23Solver);
       REQUIRE(conv_rate + tol > 3.0);
    }
 
    SECTION("SDIRK34Solver")
    {
       std::cout <<"\nTesting SDIRK34Solver" << std::endl;
-      int conv_rate = check.order(new SDIRK34Solver);
+      double conv_rate = check.order(new SDIRK34Solver);
       REQUIRE(conv_rate + tol > 4.0);
    }
 
@@ -255,28 +255,28 @@ TEST_CASE("First order ODE methods",
    SECTION("GeneralizedAlphaSolver(1.0)")
    {
       std::cout <<"\nTesting GeneralizedAlphaSolver(1.0)" << std::endl;
-      int conv_rate = check.order(new GeneralizedAlphaSolver(1.0));
+      double conv_rate = check.order(new GeneralizedAlphaSolver(1.0));
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("GeneralizedAlphaSolver(0.5)")
    {
       std::cout <<"\nTesting GeneralizedAlphaSolver(0.5)" << std::endl;
-      int conv_rate = check.order(new GeneralizedAlphaSolver(0.5));
+      double conv_rate = check.order(new GeneralizedAlphaSolver(0.5));
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("GeneralizedAlphaSolver(0.5) - restart")
    {
       std::cout <<"\nTesting GeneralizedAlphaSolver(0.5) - restart" << std::endl;
-      int conv_rate = check.order(new GeneralizedAlphaSolver(0.5), true);
+      double conv_rate = check.order(new GeneralizedAlphaSolver(0.5), true);
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("GeneralizedAlphaSolver(0.0)")
    {
       std::cout <<"\nTesting GeneralizedAlphaSolver(0.0)" << std::endl;
-      int conv_rate = check.order(new GeneralizedAlphaSolver(0.0));
+      double conv_rate = check.order(new GeneralizedAlphaSolver(0.0));
       REQUIRE(conv_rate + tol > 2.0);
    }
 
@@ -284,56 +284,56 @@ TEST_CASE("First order ODE methods",
    SECTION("AB1Solver()")
    {
       std::cout <<"\nTesting AB1Solver()" << std::endl;
-      int conv_rate = check.order(new AB1Solver());
+      double conv_rate = check.order(new AB1Solver());
       REQUIRE(conv_rate + tol > 1.0);
    }
 
    SECTION("AB1Solver() - restart")
    {
       std::cout <<"\nTesting AB1Solver() - restart" << std::endl;
-      int conv_rate = check.order(new AB1Solver(), true);
+      double conv_rate = check.order(new AB1Solver(), true);
       REQUIRE(conv_rate + tol > 1.0);
    }
 
    SECTION("AB2Solver()")
    {
       std::cout <<"\nTesting AB2Solver()" << std::endl;
-      int conv_rate = check.order(new AB2Solver());
+      double conv_rate = check.order(new AB2Solver());
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("AB2Solver() - restart")
    {
       std::cout <<"\nTesting AB2Solver() - restart" << std::endl;
-      int conv_rate = check.order(new AB2Solver(), true);
+      double conv_rate = check.order(new AB2Solver(), true);
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("AB3Solver()")
    {
       std::cout <<"\nTesting AB3Solver()" << std::endl;
-      int conv_rate = check.order(new AB3Solver());
+      double conv_rate = check.order(new AB3Solver());
       REQUIRE(conv_rate + tol > 3.0);
    }
 
    SECTION("AB4Solver()")
    {
       std::cout <<"\nTesting AB4Solver()" << std::endl;
-      int conv_rate = check.order(new AB4Solver());
+      double conv_rate = check.order(new AB4Solver());
       REQUIRE(conv_rate + tol > 4.0);
    }
 
    SECTION("AB5Solver()")
    {
       std::cout <<"\nTesting AB5Solver()" << std::endl;
-      int conv_rate = check.order(new AB5Solver());
+      double conv_rate = check.order(new AB5Solver());
       REQUIRE(conv_rate + tol > 5.0);
    }
 
    SECTION("AB5Solver() - restart")
    {
       std::cout <<"\nTesting AB5Solver() - restart" << std::endl;
-      int conv_rate = check.order(new AB5Solver(), true);
+      double conv_rate = check.order(new AB5Solver(), true);
       REQUIRE(conv_rate + tol > 5.0);
    }
 
@@ -341,56 +341,56 @@ TEST_CASE("First order ODE methods",
    SECTION("AM0Solver()")
    {
       std::cout <<"\nTesting AM0Solver()" << std::endl;
-      int conv_rate = check.order(new AM0Solver());
+      double conv_rate = check.order(new AM0Solver());
       REQUIRE(conv_rate + tol > 1.0);
    }
 
    SECTION("AM1Solver()")
    {
       std::cout <<"\nTesting AM1Solver()" << std::endl;
-      int conv_rate = check.order(new AM1Solver());
+      double conv_rate = check.order(new AM1Solver());
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("AM1Solver() - restart")
    {
       std::cout <<"\nTesting AM1Solver() - restart" << std::endl;
-      int conv_rate = check.order(new AM1Solver(), true);
+      double conv_rate = check.order(new AM1Solver(), true);
       REQUIRE(conv_rate + tol > 2.0);
    }
 
    SECTION("AM2Solver()")
    {
       std::cout <<"\nTesting AM2Solver()" << std::endl;
-      int conv_rate = check.order(new AM2Solver());
+      double conv_rate = check.order(new AM2Solver());
       REQUIRE(conv_rate + tol > 3.0);
    }
 
    SECTION("AM2Solver() - restart")
    {
       std::cout <<"\nTesting AM2Solver() - restart" << std::endl;
-      int conv_rate = check.order(new AM2Solver(), true);
+      double conv_rate = check.order(new AM2Solver(), true);
       REQUIRE(conv_rate + tol > 1.0);
    }
 
    SECTION("AM3Solver()")
    {
       std::cout <<"\nTesting AM3Solver()" << std::endl;
-      int conv_rate = check.order(new AM3Solver());
+      double conv_rate = check.order(new AM3Solver());
       REQUIRE(conv_rate + tol > 4.0);
    }
 
    SECTION("AM4Solver()")
    {
       std::cout <<"\nTesting AM4Solver()" << std::endl;
-      int conv_rate = check.order(new AM4Solver());
+      double conv_rate = check.order(new AM4Solver());
       REQUIRE(conv_rate + tol > 5.0);
    }
 
    SECTION("AM4Solver() - restart")
    {
       std::cout <<"\nTesting AM4Solver() - restart" << std::endl;
-      int conv_rate = check.order(new AM4Solver(),true);
+      double conv_rate = check.order(new AM4Solver(),true);
       REQUIRE(conv_rate + tol > 5.0);
    }
 }
